test(calendar): cover the Day Ticker's kind names, not just the map
Both surfaces publish the same `calendar.map.label` group name, so a test
that does not force the viewport exercises only the Calendar Map — and
`/simplify` caught the ticker being fed the wrong grid once already. Each
test now pins which branch it rendered by its cell count (a month grid vs
exactly TICKER_DAYS columns), so the two cannot silently collapse into one.

Also switches the kind list to `Intl.ListFormat` `style: 'long'`: the string
is an accessible name read aloud, where "Industry jobs and Planets" is a
sentence and the short form's "Industry jobs, & Planets" is not.

+++ src/features/character/dayLoadLabel.ts
@@ -27,7 +27,9 @@ export function useDayLoadLabel(): (date: Date, load: DayLoad | undefined) => st
       day: 'numeric',
       month: 'long',
     });
-    const kindList = new Intl.ListFormat(i18n.language, { style: 'short', type: 'conjunction' });
+    // `long` rather than `short`: this is read aloud, and "Industry jobs and
+    // Planets" is a sentence where "Industry jobs, & Planets" is not.
+    const kindList = new Intl.ListFormat(i18n.language, { style: 'long', type: 'conjunction' });
 
     return (date, load) =>
       load
